Design : PostView 다지인 변경, 관련 백엔드 코드 수정

+++ src/main/java/com/hobbyproject/controller/CommentRestController.java
@@ -29,8 +29,12 @@ public class CommentRestController {
     }
 
     @DeleteMapping("/post/{postId}/comment/{commentId}")
-    public void deleteComment(@PathVariable("postId") Long postId,@PathVariable("commentId") Long commentId){
-        commentService.deleteComment(commentId);
+    public void deleteComment(@PathVariable("postId") Long postId,@PathVariable("commentId") Long commentId,HttpSession session){
+        Member member = (Member) session.getAttribute("memberId");
+
+        if (commentService.isCommentOwner(commentId, member)) {
+            commentService.deleteComment(commentId);
+        }
     }
 
 }
